Notification schedule implemented
