+++ src/xcode/ENA/ENA/Source/Extensions/UNUserNotificationCenter+DeadManSwitch.swift
@@ -73,45 +73,3 @@ extension UNUserNotificationCenter {
 		removePendingNotificationRequests(withIdentifiers: [UNUserNotificationCenter.deadmanNotificationIdentifier])
 	}
 }
-
-// (pum)
-extension UNUserNotificationCenter {
-	
-	// MARK: - Internal
-	
-	/// Schedule a local notification to fire 2 hours if users has not warn to others. Scedule one more notification after 4 hours if still user has not share the keys.
-	func scheduleWarnOthersNotification(timerOne: Int, timerTwo: Int) {
-		
-		//Setup the notifications
-		let content = UNMutableNotificationContent()
-		content.title = AppStrings.WarnOthersNotification.title
-		content.body = AppStrings.WarnOthersNotification.description
-		content.sound = .default
-		
-		let trigger = UNTimeIntervalNotificationTrigger(
-			timeInterval: 2 * 60 * 60, //7200 seconds
-			repeats: false
-		)
-		
-		let request = UNNotificationRequest(
-			identifier: UNUserNotificationCenter.warnOthersNotificationIdentifier,
-			content: content,
-			trigger: trigger
-		)
-		
-		self.add(request) { error in
-			if error != nil {
-				Log.error("Warn others notification could not be scheduled.")
-			}
-		}
-	}
-	
-	// MARK: - Private
-	
-	private static let warnOthersNotificationIdentifier = (Bundle.main.bundleIdentifier ?? "") + ".notifications.cwa-warnOthers"
-	
-	/// Cancels the Warn Others Notificatoin
-	private func cancelWarnOthersNotification() {
-		removePendingNotificationRequests(withIdentifiers: [UNUserNotificationCenter.warnOthersNotificationIdentifier])
-	}
-}
